+++ scripts/iptables.sh
@@ -55,8 +55,8 @@ enable_ipv6() {
 
   # sysctl -w net.ipv6.conf.all.accept_ra=2
   # sysctl -w net.ipv6.conf.wlan0.accept_ra=2
-  # sysctl -w net.ipv6.conf.all.disable_ipv6=0
-  # sysctl -w net.ipv6.conf.default.disable_ipv6=0
+  sysctl -w net.ipv6.conf.all.disable_ipv6=0
+  sysctl -w net.ipv6.conf.default.disable_ipv6=0
   # sysctl -w net.ipv6.conf.wlan0.disable_ipv6=0
 }
 
@@ -69,10 +69,14 @@ disable_ipv6() {
 
   # sysctl -w net.ipv6.conf.all.accept_ra=0
   # sysctl -w net.ipv6.conf.wlan0.accept_ra=0
-  # sysctl -w net.ipv6.conf.all.disable_ipv6=1
-  # sysctl -w net.ipv6.conf.default.disable_ipv6=1
+  sysctl -w net.ipv6.conf.all.disable_ipv6=1
+  sysctl -w net.ipv6.conf.default.disable_ipv6=1
   # sysctl -w net.ipv6.conf.wlan0.disable_ipv6=1
 }
+
+if [ "$manual" = true ]; then
+  exit 0
+fi
 
 case "$1" in
 enable)