+++ changelog.md
@@ -1,5 +1,9 @@
 # ChangeLog
-- 修复关闭模块时开机仍然启动的问题
-- fix the problem that the module is still started when the module is closed
+- 修复manual选项不生效的问题
+- fix manual option not work
+- 更改了模块更新安装时的备份方式
+- change the backup method when updating the module
 - 更新了 README.md
 - update README.md
+- 更改了默认DNS查询方式为parallel，提高了查询速度
+- change the default DNS query method to parallel, improve the query speed